refactor(export): trim the PR to what the no-gather writer needs
Remove modelopt/torch/export/_export_common.py. It was a new module, license header and
cross-module dependency created to hold one 15-line shard-size parser, imported by exactly one
file, and carrying an unused `import torch`. Its docstring justified itself as a home for helpers
"used by more than one export module (e.g. moe_utils and distribute)" -- moe_utils never imported
it. _size_to_bytes moves into distribute.py, its only consumer.

Correct the changelog entry for this feature. It described the writer as going "through torch DCP
... consolidating them in parallel", which stopped being true when the DCP path was removed, and
attributed the large-MoE failure to "the rank-0 host-RAM ceiling". That attribution is wrong and
would have sent readers to the wrong subsystem: the gather path's Qwen3-235B-A22B failure is a GPU
OOM in _process_quantized_modules, materializing every expert on every rank during the quantizer
fold, before any writer runs. The entry now describes the actual writer and cites measurements from
16 GPUs / 4 nodes -- 405B FP8 at 149s against 307s with peak host RSS 65GB against 430GB and a
tensor-for-tensor identical checkpoint, and 235B FP8 which the gather path cannot export at all.

Scope the stale-fused-expert changelog entry to the no-gather path. The gather path still writes
the unquantized mlp.experts.* tensors alongside the quantized per-expert ones; only this writer
avoids them, and the entry read as a general fix.

GPU export suite 134 passed / 1 skipped, unchanged.
