+++ CHANGELOG.rst
@@ -22,7 +22,7 @@ Changelog
 - Add ``modelopt.torch.utils.mlflow.MlflowRunLogger`` for recording a script run on an MLflow tracking server: the invocation, the ModelOpt version, the run log (captured by teeing ``stdout``/``stderr``) and any caller-supplied artifacts, with configuration as searchable params. ``mlflow`` is an optional dependency, imported only when tracking is enabled.
 - Add ``--mlflow <tracking-uri>`` to ``examples/hf_ptq/hf_ptq.py`` (MLflow's own ``MLFLOW_TRACKING_URI`` is honoured too). A tracked run records the invocation, the resolved recipe (``$import``\ s expanded), the run log and the quantization summaries, with every command-line argument as a searchable param; failed runs are recorded with their traceback. The experiment defaults to ``$USER/hf_ptq/<checkpoint basename>-<recipe name or --qformat>`` and can be overridden with ``--mlflow_experiment`` / ``--mlflow_run_name``.
 - Add ``--mlflow <tracking-uri>`` to ``examples/vllm_serve/vllm_serve_fakequant.py`` (MLflow's own ``MLFLOW_TRACKING_URI`` is honoured too), so a fake-quant serve records what it quantized and an evaluation of that endpoint can be traced back to a recipe. A tracked run uploads the launcher command, the resolved ``RECIPE_PATH`` (or the merged ``QUANT_CFG``/``KV_QUANT_CFG`` when presets are used), the worker log and the quantizer summary; the experiment defaults to ``$USER/vllm_serve_fakequant/<model basename>-<recipe name or quantization config>`` and can be overridden with ``--mlflow-experiment`` / ``--mlflow-run-name``.
-- Export an FSDP2-sharded model to a HuggingFace checkpoint without gathering it: ``export_hf_checkpoint`` now detects an FSDP2 model under ``torch.distributed`` and has every rank write its own weight shards through torch DCP (``distributed_save_hf_checkpoint``), consolidating them in parallel, instead of gathering the full state dict to host RAM on rank 0. MoE experts stay fused and sharded through the quantizer fold, so each rank materializes only its local experts rather than all of them. This removes the rank-0 host-RAM ceiling that made large-MoE export OOM, and cuts export time on a 2-node Qwen3-30B-A3B FP8 run from 301s to 81s (94% of the old cost was the gather).
+- Export an FSDP2-sharded model to a HuggingFace checkpoint without gathering it: ``export_hf_checkpoint`` now detects an FSDP2 model under ``torch.distributed`` and routes it to ``distributed_save_hf_checkpoint``, which moves each tensor to a single owner rank and has every rank write its own shards in parallel, instead of reconstructing the full state dict in host RAM. MoE experts additionally stay fused and sharded through the quantizer fold, so each rank materializes only its local experts rather than all of them. Measured on 16 GPUs / 4 nodes: Llama-3.1-405B-Instruct FP8 exports in 149s against 307s for the gather path, with peak host RSS of 65 GB against 430 GB (the state-dict step alone falls from 99.7s to 1.6s) and a tensor-for-tensor identical checkpoint. Qwen3-235B-A22B FP8, which the gather path cannot export at all -- it exhausts GPU memory materializing every expert on every rank during the quantizer fold, before any writer runs -- exports in 8 minutes.
 
 **Backward Breaking Changes**
 
@@ -44,7 +44,7 @@ Changelog
 - Update HuggingFace checkpoint export to use name-based tied-weight deduplication instead of the previous address-based approach. The address-based deduplication could incorrectly drop an untied weight that happened to share memory with a tied one, producing an incomplete checkpoint (observed as a false positive on MiniMax-M2.7).
 - Fix EAGLE-3 training with context parallelism (``--cp_size > 1`` in ``examples/speculative_decoding``), which failed to start on ``accelerate >= 1.13`` and then raised ``got mixed torch.Tensor and DTensor``.
 - Polygraphy minimum dependency upgraded to ``0.53.4`` to solve ONNX AutoCast failures when marking optional graph outputs.
-- Fix HuggingFace export of an FSDP2-sharded MoE writing the model's **unquantized** weights into the quantized checkpoint. The exporter splits each fused 3-D expert weight into per-expert quantized weights, but the original BF16 ``mlp.experts.gate_up_proj`` / ``mlp.experts.down_proj`` parameters were still written alongside them -- with no scales, and under the names ``transformers`` binds directly -- so a quantized export also carried a full pre-quantization copy of every expert. On Qwen3-30B-A3B FP8 the experts are ~29B of the 30.5B parameters, so this was 58 GB of an 89 GB checkpoint that should be 31 GB.
+- Fix the no-gather FSDP2 export of an MoE writing the model's **unquantized** weights into the quantized checkpoint (the rank-0 gather path is unaffected by this change and still emits them). The exporter splits each fused 3-D expert weight into per-expert quantized weights, but the original BF16 ``mlp.experts.gate_up_proj`` / ``mlp.experts.down_proj`` parameters were still written alongside them -- with no scales, and under the names ``transformers`` binds directly -- so a quantized export also carried a full pre-quantization copy of every expert. On Qwen3-30B-A3B FP8 the experts are ~29B of the 30.5B parameters, so this was 58 GB of an 89 GB checkpoint that should be 31 GB.
 
 0.46 (2026-08-17)
 ^^^^^^^^^^^^^^^^^
